refactor(app): share the two pure ends of the execute pipeline
The History run view's send path was a ~130-line copy of the builder's
own `handleExecute`. The auth *rules* were already shared, but the
surrounding pipeline was not: the four-branch body construction and the
~20-field result-to-ResponseState mapping were duplicated verbatim, and
had already drifted cosmetically - the copy dropped the comments
explaining why status 0 must not default to 200 and why the body checks
test for null explicitly.

Both are pure functions of their inputs, so they move to
`utils/execute-mapping.ts` and both callers use them. The explanations
move with them, where they now cover both paths.

Not moved, deliberately: auth resolution (already shared), the
script-part list (each caller builds a different one - the builder walks
the live chain, the run view replays what was recorded), and header/URL
resolution, which differs in which variables are in scope.
